requirements.txt
amqp==1.4.9
anyjson==0.3.3
apns==2.0.1
billiard==3.3.0.23
celery==3.1.18
coreapi==2.0.8
Django==1.10.2
django-chartit==0.2.7
django-cors-headers==1.2.2
django-filter==0.15.3
django-rest-swagger==2.1.0
djangorestframework==3.5.1
djangorestframework-jwt==1.8.0
itypes==1.1.0
kombu==3.0.37
Markdown==2.6.6
openapi-codec==1.1.7
psycopg2==2.5.4
PyJWT==1.4.2
pytz==2016.7
redis==2.10.3
requests==2.11.1
simplejson==3.10.0
uritemplate==3.0.0
